Remove 2 updates
